requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
docx==0.2.4
exceptiongroup==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.26.5
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pypdf==3.4.1
PyPDF2==3.0.1
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
yapf==0.32.0
yarl==1.8.2
